Search the magic number 1f8b0800 starting from the ramdisk page.
Imgdiff tool searches the magic number 1f8b0800 to find gzipped block.
But the previous algorithm did not handle the case when that byte
sequence appears spuriously in the file.
To avoid this, this commit tries to match the magic number starting
from the ramdisk pages (skipping the kernel pages).

Change-Id: Id7e8b08884e286293ce406d0efddf0a9dc319a8f